requirements.txt
absl-py==2.3.1
astunparse==1.6.3
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
contourpy==1.3.3
cuda-bindings==12.9.3
cuda-core==0.3.2
cuda-pathfinder==1.3.1
cuda-python==12.9.3
cuda-toolkit==12.9.1
cudf-cu12==25.10.0
cupy-cuda12x==13.6.0
cycler==0.12.1
dask==2025.9.1
dask-cudf-cu12==25.10.0
distributed==2025.9.1
et_xmlfile==2.0.0
fastrlock==0.8.3
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.1
h5py==3.15.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
libcudf-cu12==25.10.0
libkvikio-cu12==25.10.0
librmm-cu12==25.10.0
llvmlite==0.44.0
locket==1.0.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpld3==0.5.11
msgpack==1.1.2
namex==0.1.0
narwhals==2.8.0
numba==0.61.2
numba-cuda==0.19.1
numpy==2.2.6
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cccl-cu12==12.9.27
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.14.0.64
nvidia-cufft-cu12==11.4.1.4
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.28.3
nvidia-nvjitlink-cu12==12.9.86
nvtx==0.2.13
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
partd==1.4.2
patsy==1.0.1
pillow==12.0.0
plotly==6.3.1
protobuf==6.33.0
psutil==7.1.1
pyarrow==21.0.0
Pygments==2.19.2
pylibcudf-cu12==25.10.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
rapids-dask-dependency==25.10.0
rapids-logger==0.1.19
requests==2.32.5
rich==14.2.0
rmm-cu12==25.10.0
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sortedcontainers==2.4.0
statsmodels==0.14.5
tblib==3.1.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow_keras==0.1
termcolor==3.1.0
threadpoolctl==3.6.0
toolz==1.1.0
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
xgboost==3.0.5
zict==3.0.0
